Implemented the barebones of verification
Implemented the barebones of verification: Now generates a unique
"verification ID" for each user on signup. There is then a verify.php
page they can be directed to which verifies the user.

Still need to generate the email.

NB: The verification ID is 30 characters long, so there are 30^64
possible combinations, which means it's somewhat secure against
bruteforce attacks. We'd be DDoS'd before it became a problem.